set compiler option "treat warnings as errors" - to prevent new warnings for wiretap
remove all compiler warnings:
a) prevent wrong malloc/free definitions by lex/yacc generated files
b) add int/time_t casts - MSVC2005 is more "sensitive" about this than MSVC6

svn path=/trunk/; revision=21078

+++ wiretap/k12.c
@@ -208,7 +208,7 @@ static gint get_record(guint8** bufferp, FILE* fh, gint64 file_offset) {
 	guint8* writep;
 	
 	/* where the next unknown 0x10 bytes are stuffed to the file */
-	gint64 junky_offset = 0x2000 - ( (file_offset - 0x200) % 0x2000 );
+	gint junky_offset = 0x2000 - (gint) ( (file_offset - 0x200) % 0x2000 );
 	
 	K12_DBG(6,("get_record: ENTER: junky_offset=%lld, file_offset=%lld",junky_offset,file_offset));
